# docs/diploma-assets/OPENAPI.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fastik API",
    "version": "0.4.0",
    "description": "REST API for the Fastik freelance marketplace."
  },
  "servers": [
    {
      "url": "http://localhost:4200/api",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "AuthUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "displayName": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "customer",
              "performer",
              "admin"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "blocked"
            ]
          },
          "emailVerified": {
            "type": "boolean"
          },
          "lastLoginAt": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/AuthUser"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a customer or performer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password",
                  "displayName",
                  "role"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  },
                  "displayName": {
                    "type": "string",
                    "minLength": 2
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "customer",
                      "performer"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registered user with access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error"
          },
          "409": {
            "description": "Email already exists"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login with email and password",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated user with access token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          },
          "403": {
            "description": "Blocked user"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get current authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/admin/overview": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Get admin operational overview",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Admin stats, permissions and recent actions"
          },
          "403": {
            "description": "Admin role required"
          }
        }
      }
    },
    "/admin/users": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List platform users for admins",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User list with balances and role state"
          },
          "403": {
            "description": "Admin role required"
          }
        }
      }
    },
    "/admin/users/{id}/status": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "Update user status",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated user"
          },
          "403": {
            "description": "Admin role required"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/admin/disputes": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List open order disputes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Disputed orders"
          },
          "403": {
            "description": "Admin role required"
          }
        }
      }
    },
    "/admin/disputes/{id}/resolve": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Resolve dispute through guarantor hold",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated dispute queue and resolution result"
          },
          "403": {
            "description": "Admin role required"
          },
          "409": {
            "description": "Dispute already resolved or invalid escrow state"
          }
        }
      }
    },
    "/admin/moderation/jobs": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List jobs for moderation",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Jobs ordered by moderation priority"
          },
          "403": {
            "description": "Admin role required"
          }
        }
      }
    },
    "/admin/moderation/jobs/{id}": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Approve or reject a job",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Moderated job"
          },
          "403": {
            "description": "Admin role required"
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
    "/admin/interviews": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List performer Elite interview states",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Performer interview queue and latest decisions"
          },
          "403": {
            "description": "Admin role required"
          }
        }
      }
    },
    "/admin/interviews/{id}/decision": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Save Elite HR interview decision for performer",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated interview queue after RPG progress recalculation"
          },
          "403": {
            "description": "Admin role required"
          },
          "404": {
            "description": "Performer not found"
          }
        }
      }
    },
    "/admin/audit-log": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "List admin audit events",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Recent admin actions"
          },
          "403": {
            "description": "Admin role required"
          }
        }
      }
    },
...
    "/marketplace/jobs": {
      "get": {
        "tags": [
          "Marketplace"
        ],
        "summary": "Get published jobs",
        "responses": {
          "200": {
            "description": "Job list"
          }
        }
      },
      "post": {
        "tags": [
          "Marketplace"
        ],
        "summary": "Create a job as customer",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Created job detail"
          },
          "403": {
            "description": "Only customers can create jobs"
          }
        }
      }
    },
    "/marketplace/jobs/{id}": {
      "get": {
        "tags": [
          "Marketplace"
        ],
        "summary": "Get job detail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job detail"
          },
          "404": {
            "description": "Job not found"
          }
        }
      }
    },
    "/marketplace/jobs/{id}/applications": {
      "post": {
        "tags": [
          "Marketplace"
        ],
        "summary": "Apply to a job as performer",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Updated job detail"
          },
          "403": {
            "description": "Only performers can apply"
          },
          "409": {
            "description": "Application conflict"
          }
        }
      }
    },
    "/marketplace/jobs/{id}/invites": {
      "post": {
        "tags": [
          "Marketplace"
        ],
        "summary": "Invite a performer to a customer job",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Updated job detail with invitation"
          },
          "403": {
            "description": "Only the job owner can invite performers"
          },
          "404": {
            "description": "Job or performer not found"
          },
          "409": {
            "description": "Invite conflict or job is not published"
          }
        }
      }
    },
    "/marketplace/jobs/{jobId}/applications/{applicationId}/select": {
      "post": {
        "tags": [
          "Marketplace"
        ],
        "summary": "Select performer application",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Job moved to in progress"
          },
          "403": {
            "description": "Only owner or manager can select"
          },
          "404": {
            "description": "Job or application not found"
          }
        }
      }
    },
...
  }
}